+++ src/main/java/bumblebee/xchangepass/global/security/refresh/RefreshToken.java
@@ -1,4 +1,4 @@
-package bumblebee.xchangepass.global.security.v1.refresh;
+package bumblebee.xchangepass.global.security.refresh;
 
 import bumblebee.xchangepass.global.error.ErrorCode;
 import lombok.AccessLevel;
