+++ src/main/java/bumblebee/xchangepass/global/security/refresh/dto/RefreshTokenRequest.java
@@ -1,0 +1,5 @@
+package bumblebee.xchangepass.global.security.refresh.dto;
+
+public record RefreshTokenRequest(String refreshToken) {
+
+}
